feat : 리뷰 검색 결과 컴포넌트 재구성
